계정 삭제기능 구현 완료

+++ src/views/signout/signout.js
@@ -33,5 +33,12 @@ async function signout() {
   const password = passwordInput.value;
   const data = { password };
 
-  await Api.delete("/api/userlist", sessionStorage.getItem("id"), data);
+  if (confirm("정말 삭제하시겠습니까?") == true) {
+    await Api.delete("/api/userlist", sessionStorage.getItem("id"), data);
+    sessionStorage.clear();
+    alert("계정이 삭제되었습니다.");
+    window.location.href = "/";
+  } else {
+    return false;
+  }
 }
